+++ cmp_internal_ui/src/features/consent/index.ts
@@ -6,6 +6,7 @@
  */
 
 export * from "@/features/consent/api";
+export * from "@/features/consent/link-url";
 export * from "@/features/consent/queries";
 export * from "@/features/consent/mutations";
 export * from "@/features/consent/components";